requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.41.1
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
idna==3.10
importlib_metadata==7.2.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.6.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow>=10.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
streamlit==1.30.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0
matplotlib==3.8.0
PyPDF2==3.0.1

